requirements.txt
beautifulsoup4==4.12.2
Flask==2.2.2
Flask_Cors==4.0.0
mysql-connector-python
nltk==3.7
openai==0.27.8
pandas
Requests==2.31.0
scikit_learn==1.1.2
gunicorn
